refactor: moved filtered event function to a separate file
